+++ Libplanet.Net.Tests/Transports/BoundPeerExtensionsTest.cs
@@ -23,10 +23,8 @@ namespace Libplanet.Net.Tests.Transports
             NetMQConfig.Cleanup(false);
         }
 
-        [Theory(Timeout = 60 * 1000)]
-        [InlineData(SwarmOptions.TransportType.NetMQTransport)]
-        [InlineData(SwarmOptions.TransportType.TcpTransport)]
-        public async Task QueryAppProtocolVersion(SwarmOptions.TransportType transportType)
+        [Fact(Timeout = 60 * 1000)]
+        public async Task QueryAppProtocolVersion()
         {
             var fx = new MemoryStoreFixture();
             var policy = new BlockPolicy<DumbAction>();
@@ -38,10 +36,7 @@ namespace Libplanet.Net.Tests.Transports
             string host = IPAddress.Loopback.ToString();
             int port = FreeTcpPort();
 
-            var option = new SwarmOptions
-            {
-                Type = transportType,
-            };
+            var option = new SwarmOptions();
 
             using (var swarm = new Swarm<DumbAction>(
                 blockchain,
@@ -53,15 +48,11 @@ namespace Libplanet.Net.Tests.Transports
             {
                 var peer = new BoundPeer(swarmKey.PublicKey, new DnsEndPoint(host, port));
                 // Before swarm starting...
-                await Assert.ThrowsAsync<TimeoutException>(async () =>
+                Assert.Throws<TimeoutException>(() =>
                 {
                     if (swarm.Transport is NetMQTransport)
                     {
                         peer.QueryAppProtocolVersionNetMQ(timeout: TimeSpan.FromSeconds(1));
-                    }
-                    else if (swarm.Transport is TcpTransport)
-                    {
-                        await peer.QueryAppProtocolVersionTcp(timeout: TimeSpan.FromSeconds(1));
                     }
                     else
                     {
@@ -78,11 +69,6 @@ namespace Libplanet.Net.Tests.Transports
                         receivedAPV = peer.QueryAppProtocolVersionNetMQ(
                             timeout: TimeSpan.FromSeconds(1));
                     }
-                    else if (swarm.Transport is TcpTransport)
-                    {
-                        receivedAPV = await peer.QueryAppProtocolVersionTcp(
-                            timeout: TimeSpan.FromSeconds(1));
-                    }
                     else
                     {
                         throw new XunitException(
@@ -97,10 +83,7 @@ namespace Libplanet.Net.Tests.Transports
                 }
             }
 
-            if (transportType == SwarmOptions.TransportType.NetMQTransport)
-            {
-                NetMQConfig.Cleanup(false);
-            }
+            NetMQConfig.Cleanup(false);
         }
 
         private static int FreeTcpPort()