+++ CHANGES.md
@@ -1,7 +1,7 @@
 Libplanet changelog
 ===================
 
-Version 0.39.0
+Version 0.40.0
 --------------
 
 To be released.
@@ -21,13 +21,41 @@ To be released.
 
 ### Behavioral changes
 
- -  `PolymorphicAction<T>` now allows subclasses of `T` to be declared in
-    the entry assembly of the application besides the same assembly as `T`.
-    [[#2136]]
+### Bug fixes
+
+### Dependencies
+
+### CLI tools
+
+
+Version 0.39.0
+--------------
+
+Released on July 18th, 2022.
+
+### Deprecated APIs
+
+ -  (Libplanet.Net) Removed `TcpTransport` class.  [[#2139]]
+ -  (Libplanet.Net) Removed `InvalidMagicCookieException` class.  [[#2139]]
+ -  (Libplanet.Net) Removed `SwarmOptions.TransportType` property.  [[#2139]]
+ -  (Libplanet.Node) Removed `InitConfig.TransportType` property.  [[#2139]]
+
+### Added APIs
+
+ -  Added `DuplicateActionTypeIdentifierException` class.  [[#2142]]
+
+### Behavioral changes
+
  -  (Libplanet.Net) Raised the default value for `Swarm<T>()` constructor's
     `workers` parameter to 100.  [[#2128], [#2134]]
  -  (Libplanet.Explorer) Raised the default value for `workers` console
     argument to 100.  [[#2128], [#2134]]
+ -  `PolymorphicAction<T>` now allows subclasses of `T` to be declared in
+    the entry assembly of the application besides the same assembly as `T`.
+    [[#2136], [#2142]]
+ -  `PolymorphicAction<T>` now throws `DuplicateActionTypeIdentifierException`
+    when multiple subtypes of `T` are associated with the same
+    `ActionTypeAttribute.TypeIdentifier`.  [[#2142]]
 
 ### Bug fixes
 
@@ -36,8 +64,6 @@ To be released.
  -  (Libplanet.Net) Invalid `Uri.UserInfo` with multiple colons is now
     rejected by `IceServer(Uri url)` constructor and exception is thrown.
     [[#2058], [#2116]]
-
-### Dependencies
 
 ### CLI tools
 
@@ -59,6 +85,8 @@ To be released.
 [#2130]: https://github.com/planetarium/libplanet/pull/2130
 [#2134]: https://github.com/planetarium/libplanet/pull/2134
 [#2136]: https://github.com/planetarium/libplanet/pull/2136
+[#2139]: https://github.com/planetarium/libplanet/pull/2139
+[#2142]: https://github.com/planetarium/libplanet/pull/2142
 [Cocona.Lite 2.0.0]: https://www.nuget.org/packages/Cocona.Lite/2.0.0